+++ src/main/kotlin/com/github/olivernybroe/wingidea/ide/editor/WingQuoteHandler.kt
@@ -1,0 +1,6 @@
+package com.github.olivernybroe.wingidea.ide.editor
+
+import com.github.olivernybroe.wingidea.lang.psi.WING_STRINGS
+import com.intellij.codeInsight.editorActions.SimpleTokenSetQuoteHandler
+
+class WingQuoteHandler: SimpleTokenSetQuoteHandler(WING_STRINGS)