+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for caret position in arguments
 - Added brace matching support
 - Added commenter support
+- Added quote handler
 
 ### Fixed
 - Fixed argument with argument name not being parsed correctly